pyphisher
Disclaimer 
Please note that any material or activity related to cybersecurity or hacking provided here is intended solely for educational purposes. The purpose of this material is to increase awareness and understanding of cybersecurity concepts and principles. The activities described in this material should not be attempted outside of a controlled educational setting. Any attempt to engage in illegal or unethical activities related to cybersecurity could result in serious legal and ethical consequences. It is the responsibility of individuals to use this information in a responsible and ethical manner.